Use inline test modules and move mod statements to the top
