+++ daphne/src/roles/leader.rs
@@ -1,3 +1,6 @@
+// Copyright (c) 2023 Cloudflare, Inc. All rights reserved.
+// SPDX-License-Identifier: BSD-3-Clause
+
 use std::{borrow::Cow, collections::HashMap};
 
 use async_trait::async_trait;
